==> 1.Scraping data.ipynb <==
# Script de scraping 

import requests as rq
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
from bs4 import UnicodeDammit

## 'Scrapeamos' los datos de  www.autodescuento.com, especializados en reunir información de las mejores ofertas que se vienen recogiendo de concesionarios oficiales en toda España.

### Este script almacena todas las versiones de todos los modelos a la venta en España y sus principales características. 

main_page = rq.get("http://www.autodescuento.com")
main_soup = BeautifulSoup(main_page.content, 'html.parser')
path_marcas = "http://www.autodescuento.com/coches-nuevos-"

lista_marcas = []

for marca in main_soup.find_all('h2', class_='h2'):
    
    lista_marcas.append(marca.get_text().strip().replace(' ','-'))

lista_marcas_limpia = lista_marcas[1:-2]


versiones_modelo = []

for marca in lista_marcas_limpia:
    
    second_page = rq.get(path_marcas+marca+".htm", headers={'User-Agent': 'Safari'})
    second_soup = BeautifulSoup(second_page.content, 'html.parser')
    #print('Otra marca')
    for modelo in second_soup.find('div', class_='izquierda').find_all('a', class_="coche redondas"):
        
        #print(modelo['href'])
        #model = modelo.find('p',class_='mm').getText().encode('utf-8') + "-"
        #print(type(model))
        third_page = rq.get(modelo['href'], headers={'User-Agent': 'Safari'})
        third_soup = BeautifulSoup(third_page.content, 'html.parser')
        
        if(third_soup.find('span',id='tdversiones') is not None):
            
            versiones = third_soup.find('span',id='tdversiones').find_all('a', href=True)

            for ver in versiones:

                forth_page = rq.get(ver.get('href').encode('utf-8'),headers={'User-Agent': 'Safari'})
                forth_soup = BeautifulSoup(forth_page.content, 'html.parser')
                
                #linea_modelo = model
                #print(linea_modelo)
                linea_modelo = forth_soup.find('h1').getText().encode('utf-8').replace(' - ',',')+','+ver.find('div', class_='precio').getText().encode('utf-8')
    
                for campos in forth_soup.find_all('label', class_='campos'):
        
                    linea_modelo += ","+campos.find_next('span').getText().encode('utf-8').replace(',','.')        
                
                versiones_modelo.append(linea_modelo)
                                                                   
                                                                 

### Se guarda el resultado en .csv como checkpoint para el análisis posterior.  

df = pd.DataFrame(versiones_modelo)
pd.set_option('display.max_columns', None)
df.to_csv('datos_crudo',encoding='utf-8',index=False)

### Se eliminan marcadores ' " ' para su correcta lectura con pandas posteriormente. 

!sed 's/"//g' datos_crudo > datos_limpios

### Lectura del archivo y eliminación de columnas innecesarias. 

colnames = [
    'marca',
    'modelo',
    'version',
    'precio',
    'puertas',
    'desc',
    'combustible',
    'traccion',
    'cv',
    'cambio',
    'n_marchas',
    'co2',
    'autonomia',
    'consumo',
    'airbags',
    'garantia',
    'desc4',
    'comb',
    'co2',
    'cilindrada',
    'traccion',
    'capacidad deposito',
    'cambio',
    'autonomia',
    'marchas',
    'traccion',
    'max_vel',
    '0_a_100',
    'cv',
    'consumo',
    'rpm',
    'par',
    'airbags',
    'control traccion',
    'control estabilidad',
    'estrellas NCAP',
    'desc9',
    'inmovilizador',
    'alarma',
    'cierre centralizado',
    'cierre automatico',
    'tipo llave',
    'start/stop',
    'desc16',
    'n_altavoces',
    'desc18',
    'CD/mp3/dvd',
    'cargador',
    'carroceria',
    'medidas largo/ancho/alto',
    'puertas',
    'peso',
    'desc24'
]

df_1 = pd.read_csv('datos_limpios', sep=',',names=colnames)
df_1 = df_1.drop(0)

colnames_to_delete = [
    'desc',
    'desc4',
    'comb',
    'traccion.1',
    'cambio.1',
    'autonomia.1',
    'cv.1',
    'co2.1',
    'consumo.1',
    'traccion.2',
    'inmovilizador',
    'airbags.1',
    'desc9',
    'desc16',
    'n_altavoces',
    'desc18',
    'cargador',
    'medidas largo/ancho/alto',
    'puertas.1',
    'desc24'
]

df_1 = df_1.drop(labels=colnames_to_delete,axis=1)
df_1.to_csv('Dataframe_limpio.csv', encoding='utf-8')

## Scrapeamos los mejores precios oficiales.

import requests as rq


main_page = rq.get("http://www.autodescuento.com")
main_soup = BeautifulSoup(main_page.content, 'html.parser')
path_marcas = "http://www.autodescuento.com/coches-nuevos-"

lista_marcas = []

for marca in main_soup.find_all('h2', class_='h2'):
    
    lista_marcas.append(marca.get_text().strip().replace(' ','-'))

lista_marcas_limpia = lista_marcas[1:-2]


precios_modelos = []

for marca in lista_marcas_limpia:
    
    second_page = rq.get(path_marcas+marca+".htm", headers={'User-Agent': 'Safari'})
    second_soup = BeautifulSoup(second_page.content, 'html.parser')
    
    for modelo in second_soup.find('div', class_='izquierda').find_all('a', class_="coche redondas"):

        model = modelo.find('p',class_='mm').getText().encode('utf-8') + "-"
        third_page = rq.get(modelo['href'], headers={'User-Agent': 'Safari'})
        third_soup = BeautifulSoup(third_page.content, 'html.parser')
        
        if(third_soup.find('span',id='tdversiones') is not None):

            for ver in third_soup.find('span',id='tdversiones').find_all('a', href=True):
                
                precios_modelos.append(ver.find('div', class_='precio').getText())
                
                
                



### Guardamos los datos de los precios como checkpoint. 

df_precios = pd.DataFrame({'precio':precios_modelos})
df_precios.to_csv('precios_versiones',encoding='utf-8')

## En este punto, disponemos de la base de datos de vehículos y sus características, incluidas sus mejores ofertas.
