==> 2.Scraping extra data - Maleteros.ipynb <==
# Scraping de capacidades de maleteros 

from bs4 import BeautifulSoup
import pandas as pd
import requests as rq

### Script que scrapea los datos oficiales de capacidades de maleteros de los diferentes vehículos, a través de www.medidasdecoches.com/ .  

main_url = 'http://www.medidasdecoches.com/'
main_page = rq.get(main_url)
main_soup = BeautifulSoup(main_page.content, 'html.parser')

capacidades = []

for marca in main_soup.find_all('div', class_='logos'):
    
    #print(marca.find('a').getText())
    second_page = rq.get(main_url+marca.find('a', href=True)['href'])
    second_soup = BeautifulSoup(second_page.content, 'html.parser')
    
    for modelo in second_soup.find_all('div',class_='unit'):
        
        #print(modelo.find('br').getText())
        capacidades.append(marca.find('a').getText()+':'+modelo.find('h2').getText()+':'+modelo.find('br').getText())
        #print(modelo.find('h2').getText()+':'+modelo.find('br').getText())
        #print(modelo.find('br').getText())
        

### Ordenación de los datos recopilados. 

df_capacidades = pd.DataFrame(capacidades,columns=['Marca : Modelo : Titulo : Dimensiones : Maletero'])
df_capacidades = df_capacidades['Marca : Modelo : Titulo : Dimensiones : Maletero'].apply(lambda x: (pd.Series(x.split(':'))))
df_capacidades['Marca']=df_capacidades[0]
df_capacidades['Modelo']=df_capacidades[1]
df_capacidades['Maletero']=df_capacidades[4]
df_capacidades = df_capacidades.drop([0,1,2,3,4],axis=1)
df_capacidades = df_capacidades.dropna()

def sacar_numero(capacidad_unidades):
    capacidad = str(capacidad_unidades)
    trozos = [capacidad.split(' ')]
    return trozos[0]

#df_capacidades['Maletero']=df_capacidades['Maletero'].apply()
#df_capacidades.head()

### Se guardan los datos como checkpoint. 

df_capacidades['CapacidadMaletero'] = df_capacidades['Maletero'].apply(lambda x: x.strip()[0:x.rfind(' ')])
df_capacidades = df_capacidades.drop(['Maletero'], axis=1)
df_capacidades.to_csv('maleteros',encoding='utf-8')
